minor changes on code

+++ fwork_207931106/q_2/q_22/client.c
@@ -9,6 +9,10 @@ int main(int argc, char **argv) {
     sscanf(argv[1],"%d",&server_pid);
     sscanf(argv[2],"%d",&signal_num);
     sscanf(argv[3],"%d",&signal_count);
+    if(signal_num != 2 || signal_num != 10) {
+        printf("signal_num is illegal! \n please choose 2 for SIGINT or 10 for SIGUSR1");
+        exit(1);
+    }
     int i = 0;
     while(i < signal_count) {
         kill(server_pid, signal_num);
